Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325313 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 9f47aaa46710d632d472e89a419a3af42a65008b

+++ src/components/org/apache/jmeter/visualizers/Graph.java
@@ -186,7 +186,7 @@ public class Graph extends JComponent implements Scrollable, Clearable
          repaint();
          return;
       }
-      final int xPos = (int) model.getCount();
+      final int xPos = model.getCount();
 
       SwingUtilities.invokeLater(new Runnable()
       {
